backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
bcrypt==5.0.0
click==8.3.1
colorama==0.4.6
ecdsa==0.19.1
fastapi==0.129.0
greenlet==3.3.1
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2-binary==2.9.11
pyasn1==0.6.2
pydantic[email]==2.12.5
pydantic_core==2.41.5
pydantic-settings==2.1.0
python-dotenv==1.2.1
python-multipart==0.0.9
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.46
starlette==0.52.1
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.40.0